parsing and saving of memory and cpu usage done

+++ in.BBAT.TestRunner/src/in/BBAT/TestRunner/Listener/ICpuUsageListener.java
@@ -2,7 +2,7 @@ package in.BBAT.TestRunner.Listener;
 
 public interface ICpuUsageListener {
 
-	void cpuUsage(int percentUsage,long time);
+	void cpuUsage(double percentUsage,long time);
 	
 	/**
 	 *
